bpf: use dev_scratch for SK_SKB type programs metadata
It is often useful to have a per skb scratch pad to use with
bpf to store metadata across tail calls and/or BPF programs.

However, SK_SKB programs can not use skb->mark region because
at the socket layer it used to hold a reserved_tail field for
the stream alloc code paths. This patch exposes the dev_scratch
field to BPF programs which has a similar usage to skb mark but
can be used without conflict by SK_SKB.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>