bpf: add sch_bpf qdisc
This work adds a class-less, bpf based qdisc, which can be used
on ingress or egress. It sets up mini-qdiscs and plugs into the
existing sch_handle_ingress() and sch_handle_egress() points
similarly to other pseudo qdiscs such as ingress and clsact.

The bpf based qdisc can be used as an alternative option to run
in software or offloaded through the same underlying infrastructure
via ndo_setup_tc() driver callback. The offloading model simplifies
a bit as well for just having to deal with a single program; support
for nfp has been implemented to reuse its JIT. It also simplifies
that once programs are offloaded that no extra code is needed in
software fast-path to handle the skip for non-offloaded programs.

There's also no extra overhead in fast-path for other qdiscs such
as ingress and clsact qdiscs. The programs piggy-back on the
mini-qdisc swap that we have in sch_handle_ingress() and
sch_handle_egress() fast-path and allows for reusing bpf object
files written for cls_bpf without any changes needed. The ingress
and egress callbacks have been split such that each side only
needs to deal with the minimum amount of work necessary to propely
invoke their bpf program.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>