bpf: sockmap, add msg_cork_bytes() helper
In the case where we want to apply a verdict to multiple sendmsg
calls or to a subset of the BPF program may use msg_apply_bytes().
This works well and is sufficient for streaming cases. However,
another use case exists where the BPF user will want to block
messages until the entire msg is validated and marked "correct".

The current infrastructure can not support this because a user
may, in the extreme case, call sendmsg with 1-byte msg segments.
To allow BPF programs to support collecting multiple msg calls
into a single BPF data segment add a new helper call,
msg_cork_bytes().

Signed-off-by: John Fastabend <john.fastabend@gmail.com>