bpf: tls convert plaintext data into sk_msg_buff to use sk_msg progs
We would like to use sk_msg BPF programs with KTLS but, BPF uses the
sk_msg_buff structure as input into the BPF handlers. However, we
notice that many of the fields in the TLS sw structure are tracking
duplicate fields in the sk_msg_buff. So to make the KTLS and sk_msg
coexist embed the sk_msg_buff structure into the ktls sw struct.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>