bpf, sockmap: ingress redirect with encryption header breaks receive thread
When using BPF_SK_SKB_STREAM_VERDICT to redirect a skb into the ingress
path of a sock the first 13B contain the protocol header. We could simply
strip it but this could break other use cases that require it. And further
encoding this policy into the datapath is bound to be problematic to cover
every corner case where folks want to strip it or not.

So in order to fix this case add skb_adjust_room() so users can strip the
header. After this the user can strip the header and an unmodified receiver
thread will work correctly when data is redirected into the ingress path
of a sock.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>