bpf, sockmap: Pull orphan out and use simpler design using skb_kfree
Original design used skb_orphan() which called the destructor hooks
which would return memory to the socket. We did this early in sk_skb
handling to return the memory early and then charged it to the redirect
socket immediately after running the BPF program and getting its
verdict.

But, this design caused a fundamental problem. Namely, what should we do
if we redirect to a socket that has already reached its socket memory
limits. For proxy use cases where the network admin can tune memory
limits we punted on this problem and told folks to simply make your memory
limits high enough to handle your workload. This is not a really good
answer. When deploying into environments where we expect this to be
transparent its no longer the case because we need to tune parameters.
In fact its really only viable in cases where we have fine grained
control over the application. For example a proxy redirecting from an
ingress socket to an egress socket.

Previous patches fix the case where we redirect into a socket and
encounter a memory limit on the socket. Now we can retry from the
workqueue with an EAGAIN error. This should never drop data unless the
pipe breaks.

But, in current state we have a small window where memory is not charged
to either the old socket or the new socket. Having unaccounted memory
is not great. The fix turns out to be straight-forwared. In this patch
we remove the skb_orphan() calls and use the skb_kfree() to free the
memory. This way memory is accountable to the original socket until the
actual send/recv of the data is complete. Removing the window where
the memory is not being accounted for.

This also puts a limit on the backlog queue because all skb data here
is accounted for with rmem socket limits.

Fixes: 604326b41a6fb ("bpf, sockmap: convert to generic sk_msg interface")
Signed-off-by: John Fastabend <john.fastabend@gmail.com>